Move into nested folders based on endpoint
